Implemented get specific note functionality w/update functionality

+++ frontend/renderer/components/sidebar.jsx
@@ -7,7 +7,7 @@ import { AiFillFolderOpen } from 'react-icons/ai';
 import { AiOutlineClockCircle } from 'react-icons/ai'
 import { FaTrash } from "react-icons/fa";
 
-const Sidebar = () => {
+const Sidebar = ({setSelectedNote}) => {
   const [notes, setNotes] = useState([]);
   const [filtered, setFiltered] = useState(notes);
   const inputRef = useRef(null);
@@ -38,11 +38,29 @@ const Sidebar = () => {
   }
 
   const deleteNote = async (id) => {
-    await axios.delete(`http://localhost:8080/api/notes/${id}`); 
+    try {
+      if (!id) {
+        throw new Error("No id was passed to delete a note.");
+      }
+
+      await axios.delete(`http://localhost:8080/api/notes/${id}`); 
+    }
+    catch (error) {
+      console.error("Error: " + error);
+    }
   }
 
-  const getNote = () => {
+  const getNote = async (id) => {
+    try {
+      if (!id) {
+        throw new Error("No id was passed to retrieve a note.");
+      }
 
+      const response = await axios.get(`http://localhost:8080/api/notes/${id}`); 
+    }
+    catch (error) {
+      console.error(error);
+    }
   }
 
   const searchNote = (event) => {
@@ -92,16 +110,25 @@ const Sidebar = () => {
         <div className="my-4">
           {filtered.sort((a, b) => b.updatedAt - a.updatedAt).map((note) => {
             return (
-              <div className="my-2 p-2 border rounded-lg border-border bg-editor" key={note.noteId}>
-                <h3 className="text-muted font-serif">{note.title}</h3>
-                <div className="text-text flex items-center gap-2 mt-3 text-xs text-muted-foreground/70">
-                  <AiOutlineClockCircle/>
-                  <span className="text-sm text-text font-serif"> Last edited: {parseTimestamp(note.updatedAt)}</span>
-                  <a onClick={() => deleteNote(note.noteId)}>
-                    <FaTrash className="hover:text-red-400"/>
-                  </a>
+              <a onClick={() => {
+                if (notes.length === 0) {
+                  setSelectedNote(null);
+                }
+                else {
+                  setSelectedNote(note.noteId)
+                }
+              }}>
+                <div className="my-2 p-2 border rounded-lg border-border bg-editor" key={note.noteId}>
+                  <h3 className="text-muted font-serif">{note.title}</h3>
+                  <div className="text-text flex items-center gap-2 mt-3 text-xs text-muted-foreground/70">
+                    <AiOutlineClockCircle/>
+                    <span className="text-sm text-text font-serif"> Last edited: {parseTimestamp(note.updatedAt)}</span>
+                    <a onClick={() => deleteNote(note.noteId)}>
+                      <FaTrash className="hover:text-red-400"/>
+                    </a>
+                  </div>
                 </div>
-              </div>
+              </a>
             )
           })}
         </div>
